requirements.txt
Django==4.0
Pillow==9.3.0
environs==9.5.0
django-tinymce==3.5.0
django-admin-sortable2==2.1.4
requests==2.28.2
django-debug-toolbar==3.2.3
urllib3==1.26.6